Initial commit for moving streams and AssetID to Mezz namespace.
Moved the files for Streams and AssetID and updated the files that use
them to reflect the move.
