+++ Mezzanine/CMakeLists.txt
@@ -80,6 +80,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}actormanager.h"
         "${${PROJECT_NAME}_SRC_DIR}areaeffect.h"
         "${${PROJECT_NAME}_SRC_DIR}areaeffectmanager.h"
+        "${${PROJECT_NAME}_SRC_DIR}assetid.h"
         "${${PROJECT_NAME}_SRC_DIR}attachable.h"
         "${${PROJECT_NAME}_SRC_DIR}axisalignedbox.h"
         "${${PROJECT_NAME}_SRC_DIR}binarybuffer.h"
@@ -90,6 +91,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}crossplatform.h"
         "${${PROJECT_NAME}_SRC_DIR}crossplatformexport.h"
         "${${PROJECT_NAME}_SRC_DIR}cubicspline.h"
+        "${${PROJECT_NAME}_SRC_DIR}datastream.h"
         "${${PROJECT_NAME}_SRC_DIR}datatypes.h"
         "${${PROJECT_NAME}_SRC_DIR}debris.h"
         "${${PROJECT_NAME}_SRC_DIR}debrismanager.h"
@@ -126,6 +128,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}eventuserinput.h"
         "${${PROJECT_NAME}_SRC_DIR}exception.h"
         "${${PROJECT_NAME}_SRC_DIR}fieldofforce.h"
+        "${${PROJECT_NAME}_SRC_DIR}filestream.h"
         "${${PROJECT_NAME}_SRC_DIR}gravityfield.h"
         "${${PROJECT_NAME}_SRC_DIR}gravitywell.h"
         "${${PROJECT_NAME}_SRC_DIR}hashedstring.h"
@@ -139,6 +142,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}managerfactory.h"
         "${${PROJECT_NAME}_SRC_DIR}matrix3x3.h"
         "${${PROJECT_NAME}_SRC_DIR}matrix4x4.h"
+        "${${PROJECT_NAME}_SRC_DIR}memorystream.h"
         "${${PROJECT_NAME}_SRC_DIR}meshterrain.h"
         "${${PROJECT_NAME}_SRC_DIR}mezzanine.h"
         "${${PROJECT_NAME}_SRC_DIR}multibodyvehicle.h"
@@ -486,10 +490,6 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}Resource/asset.h"
         "${${PROJECT_NAME}_SRC_DIR}Resource/assetgroup.h"
         "${${PROJECT_NAME}_SRC_DIR}Resource/assethandler.h"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/assetid.h"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/datastream.h"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/filestream.h"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/memorystream.h"
         "${${PROJECT_NAME}_SRC_DIR}Resource/resourceutilities.h"
         # Scripting Items
         "${${PROJECT_NAME}_SRC_DIR}Scripting/script.h"
@@ -630,6 +630,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}circle.cpp"
         "${${PROJECT_NAME}_SRC_DIR}colourvalue.cpp"
         "${${PROJECT_NAME}_SRC_DIR}crossplatform.cpp"
+        "${${PROJECT_NAME}_SRC_DIR}datastream.cpp"
         "${${PROJECT_NAME}_SRC_DIR}debris.cpp"
         "${${PROJECT_NAME}_SRC_DIR}debrismanager.cpp"
         "${${PROJECT_NAME}_SRC_DIR}entresol.cpp"
@@ -645,6 +646,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}eventuserinput.cpp"
         "${${PROJECT_NAME}_SRC_DIR}exception.cpp"
         "${${PROJECT_NAME}_SRC_DIR}fieldofforce.cpp"
+        "${${PROJECT_NAME}_SRC_DIR}filestream.cpp"
         "${${PROJECT_NAME}_SRC_DIR}gravityfield.cpp"
         "${${PROJECT_NAME}_SRC_DIR}gravitywell.cpp"
         "${${PROJECT_NAME}_SRC_DIR}hashedstring.cpp"
@@ -655,6 +657,7 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}managerbase.cpp"
         "${${PROJECT_NAME}_SRC_DIR}matrix3x3.cpp"
         "${${PROJECT_NAME}_SRC_DIR}matrix4x4.cpp"
+        "${${PROJECT_NAME}_SRC_DIR}memorystream.cpp"
         "${${PROJECT_NAME}_SRC_DIR}meshterrain.cpp"
         "${${PROJECT_NAME}_SRC_DIR}multibodyvehicle.cpp"
         "${${PROJECT_NAME}_SRC_DIR}objectsettings.cpp"
@@ -948,9 +951,6 @@ set(${PROJECT_NAME}_sources
         "${${PROJECT_NAME}_SRC_DIR}Resource/asset.cpp"
         "${${PROJECT_NAME}_SRC_DIR}Resource/assetgroup.cpp"
         "${${PROJECT_NAME}_SRC_DIR}Resource/assethandler.cpp"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/datastream.cpp"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/filestream.cpp"
-        "${${PROJECT_NAME}_SRC_DIR}Resource/memorystream.cpp"
         "${${PROJECT_NAME}_SRC_DIR}Resource/resourceutilities.cpp"
         # Scripting Items
         "${${PROJECT_NAME}_SRC_DIR}Scripting/scriptingmanager.cpp"
