+++ Mezzanine/src/Audio/OALS/oalssoundproxy.h
@@ -206,10 +206,10 @@ namespace Mezzanine
                 virtual void SetPitch(const Real Pitch);
                 /// @copydoc iSound::GetPitch() const
                 virtual Real GetPitch() const;
-                /// @copydoc iSound::SetStream(Resource::DataStreamPtr, const Audio::Encoding)
-                virtual void SetStream(Resource::DataStreamPtr Stream, const Audio::Encoding Encode);
-                /// @copydoc iSound::SetStream(const UInt16, Resource::DataStreamPtr, const Audio::Encoding)
-                virtual void SetStream(const UInt16 Type, Resource::DataStreamPtr Stream, const Audio::Encoding Encode);
+                /// @copydoc iSound::SetStream(DataStreamPtr, const Audio::Encoding)
+                virtual void SetStream(DataStreamPtr Stream, const Audio::Encoding Encode);
+                /// @copydoc iSound::SetStream(const UInt16, DataStreamPtr, const Audio::Encoding)
+                virtual void SetStream(const UInt16 Type, DataStreamPtr Stream, const Audio::Encoding Encode);
                 /// @copydoc iSound::SetStream(iDecoder*)
                 virtual void SetStream(iDecoder* Decode);
                 /// @copydoc iSound::SetStream(const UInt16, iDecoder*)
