+++ Mezzanine/src/Audio/decoderfactory.h
@@ -64,7 +64,7 @@ namespace Mezzanine
             /// @brief Creates a new @ref iDecoder for use.
             /// @param Stream A smart pointer to the stream containing audio data.
             /// @return Returns the newly created @ref iDecoder, or NULL if there was a failure.
-            virtual iDecoder* CreateDecoder(Resource::DataStreamPtr Stream) = 0;
+            virtual iDecoder* CreateDecoder(DataStreamPtr Stream) = 0;
             /// @brief Gets the encoding supported by decoders made by this factory.
             /// @return Returns an @ref Audio::Encoding value representing the encoding supported by this factory.
             virtual Audio::Encoding GetSupportedEncoding() const = 0;
