+++ Mezzanine/src/Audio/soundscapemanager.cpp
@@ -78,22 +78,22 @@ namespace Mezzanine
         SoundProxy* SoundScapeManager::CreateMusicSoundProxy()
             { return this->CreateSoundProxy(Audio::ST_Music); }
 
-        SoundProxy* SoundScapeManager::CreateAmbientSoundProxy(Resource::DataStreamPtr Stream, const Audio::Encoding Encode)
+        SoundProxy* SoundScapeManager::CreateAmbientSoundProxy(DataStreamPtr Stream, const Audio::Encoding Encode)
             { return this->CreateSoundProxy(Audio::ST_Ambient,Stream,Encode); }
-        SoundProxy* SoundScapeManager::CreateDialogSoundProxy(Resource::DataStreamPtr Stream, const Audio::Encoding Encode)
+        SoundProxy* SoundScapeManager::CreateDialogSoundProxy(DataStreamPtr Stream, const Audio::Encoding Encode)
             { return this->CreateSoundProxy(Audio::ST_Dialog,Stream,Encode); }
-        SoundProxy* SoundScapeManager::CreateEffectSoundProxy(Resource::DataStreamPtr Stream, const Audio::Encoding Encode)
+        SoundProxy* SoundScapeManager::CreateEffectSoundProxy(DataStreamPtr Stream, const Audio::Encoding Encode)
             { return this->CreateSoundProxy(Audio::ST_Effect,Stream,Encode); }
-        SoundProxy* SoundScapeManager::CreateMusicSoundProxy(Resource::DataStreamPtr Stream, const Audio::Encoding Encode)
+        SoundProxy* SoundScapeManager::CreateMusicSoundProxy(DataStreamPtr Stream, const Audio::Encoding Encode)
             { return this->CreateSoundProxy(Audio::ST_Music,Stream,Encode); }
 
-        SoundProxy* SoundScapeManager::CreateAmbientSoundProxy(Resource::DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
+        SoundProxy* SoundScapeManager::CreateAmbientSoundProxy(DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
             { return this->CreateSoundProxy(Audio::ST_Ambient,Stream,Frequency,Config); }
-        SoundProxy* SoundScapeManager::CreateDialogSoundProxy(Resource::DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
+        SoundProxy* SoundScapeManager::CreateDialogSoundProxy(DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
             { return this->CreateSoundProxy(Audio::ST_Dialog,Stream,Frequency,Config); }
-        SoundProxy* SoundScapeManager::CreateEffectSoundProxy(Resource::DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
+        SoundProxy* SoundScapeManager::CreateEffectSoundProxy(DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
             { return this->CreateSoundProxy(Audio::ST_Effect,Stream,Frequency,Config); }
-        SoundProxy* SoundScapeManager::CreateMusicSoundProxy(Resource::DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
+        SoundProxy* SoundScapeManager::CreateMusicSoundProxy(DataStreamPtr Stream, const UInt32 Frequency, const Audio::BitConfig Config)
             { return this->CreateSoundProxy(Audio::ST_Music,Stream,Frequency,Config); }
 
         SoundProxy* SoundScapeManager::CreateAmbientSoundProxy(const String& FileName, const String& Group)
