+++ Mezzanine/src/Audio/wavdecoder.h
@@ -60,7 +60,7 @@ namespace Mezzanine
         protected:
             /// @internal
             /// @brief This is a shared pointer to the stream being decoded.
-            Resource::DataStreamPtr WavStream;
+            DataStreamPtr WavStream;
             /// @internal
             /// @brief This is used to cache the total size of the stream used by this decoder.
             Integer WavStreamSize;
@@ -97,11 +97,11 @@ namespace Mezzanine
             /// @internal
             /// @brief Reads the metadata at the start of a Wav stream and saves it.
             /// @param Stream The Wav stream to read metadata from.
-            void ReadWavMetaData(Resource::DataStreamPtr Stream);
+            void ReadWavMetaData(DataStreamPtr Stream);
         public:
             /// @brief Class constructor.
             /// @param Stream The stream to decode.
-            WavDecoder(Resource::DataStreamPtr Stream);
+            WavDecoder(DataStreamPtr Stream);
             /// @brief Class destructor.
             virtual ~WavDecoder();
 
@@ -119,7 +119,7 @@ namespace Mezzanine
             /// @copydoc iDecoder::GetFrequency() const
             UInt32 GetFrequency() const;
             /// @copydoc iDecoder::GetStream() const
-            Resource::DataStreamPtr GetStream() const;
+            DataStreamPtr GetStream() const;
             /// @copydoc iDecoder::IsEndOfStream() const
             Boole IsEndOfStream() const;
