+++ Mezzanine/src/UI/textureatlashandler.cpp
@@ -43,8 +43,9 @@
 #include "UI/textureatlashandler.h"
 #include "UI/textureatlas.h"
 
+#include "datastream.h"
+
 #include "Resource/resourcemanager.h"
-#include "Resource/datastream.h"
 #include "XML/xml.h"
 
 namespace Mezzanine
@@ -63,7 +64,7 @@ namespace Mezzanine
         void TextureAtlasHandler::LoadAtlasFromFile(const String& Name, const String& Group)
         {
             /// @todo Update after we have refactored the resource system if needed.
-            Resource::DataStreamPtr AtlasStream = Resource::ResourceManager::GetSingletonPtr()->OpenAssetStream(Name,Group);
+            DataStreamPtr AtlasStream = Resource::ResourceManager::GetSingletonPtr()->OpenAssetStream(Name,Group);
             AtlasStream->SetStreamPosition(0);
             XML::Document AtlasDoc;
             AtlasDoc.Load( *AtlasStream.Get() );
