+++ Mezzanine/src/assetid.h
